Add files via upload
Project Objective
The best way for you to understand the concepts of an Operating System is to build an operating system and then to experiment with it to see how the OS manages resources and processes. In this project, you are asked to build a simulation of an operating system. The main focus of the project will be on building a correct architecture that simulates a real operating system. 

Description
In this project, you are asked to implement a basic interpreter. You have a text file that represents a program. When you read that text file and start executing it, it becomes a process. You are provided with 3 program files each rep- resenting a program. You are asked to create an interpreter that reads the txt files and executes their code. You are asked to implement a memory and save the processes in it. You are also asked to implement mutexes that ensure mutual exclusion over the critical resources. And finally, You are also asked to implement a scheduler that schedules the processes that we have in our system.